splited parser.go to several files as refactoring.
