+++ parser/parser_expressions.go
@@ -1,0 +1,23 @@
+package parser
+
+import (
+	"github.com/ryogrid/SamehadaDB/execution/expression"
+	"github.com/ryogrid/SamehadaDB/types"
+)
+
+type BinaryOpExpression struct {
+	LogicalOperationType_    expression.LogicalOpType
+	ComparisonOperationType_ expression.ComparisonType
+	Left                     interface{}
+	Right                    interface{}
+}
+
+type SetExpression struct {
+	ColName_     *string
+	UpdateValue_ *types.Value
+}
+
+type ColDefExpression struct {
+	ColName_ *string
+	ColType_ *types.TypeID
+}
