+++ parser/visitor.go
@@ -1,0 +1,10 @@
+package parser
+
+import (
+	"github.com/pingcap/parser/ast"
+)
+
+type Visitor interface {
+	Enter(n ast.Node) (node ast.Node, skipChildren bool)
+	Leave(n ast.Node) (node ast.Node, ok bool)
+}
